+++ CHANGELOG.md
@@ -6,6 +6,18 @@
   > in order to enable kube "securityContext.runAsNonRoot: true", docker image should use numeric user (UID)
 * monitor: fixed kafka high disk alert message
   > kafka disk usage uses size as threshold, alert message should convert to percentage
+* log-processor: support action log trace filter
+  > to filter out trace to reduce storage, e.g. under massive scanning
+  > under kube, e.g.
+  > - name: APP_FILTER_CONFIG
+      value: |
+      {
+      "action": {
+      "ignoreTrace": [
+      {"apps": ["website"], "errorCodes": ["REQUEST_BLOCKED"]}
+      ]
+      }
+      }
 
 ### 8.0.6 (08/16/2022 - 09/23/2022)